Remove the cache validation
The cache validation worked based on timestamps. I believe this
may be unreliable in a workflow pipeline with cached assets.

Instead, we now rely on the digest of the file. If the source file
changes then so should the digest. And therefore it would not find
a file in the cache.